ogre_helpers: refactor out non QWidget related functions
