+++ src/rviz/ogre_helpers/qt_widget_ogre_render_window.cpp
@@ -2,6 +2,9 @@
 #include "orthographic.h"
 #include "render_system.h"
 
+#include <QApplication>
+#include <QWindow>
+
 #include <OgreRoot.h>
 #include <OgreViewport.h>
 #include <OgreCamera.h>
@@ -21,215 +24,52 @@ namespace rviz
 {
 
 QtWidgetOgreRenderWindow::QtWidgetOgreRenderWindow( QWidget* parent )
-  : RenderWidget( RenderSystem::get(), parent )
-  , viewport_( nullptr )
-  , ogre_root_( RenderSystem::get()->root() )
-  , ortho_scale_( 1.0f )
-  , auto_render_( true )
-  , camera_( nullptr )
-  , overlays_enabled_( true ) // matches the default of Ogre::Viewport.
-  , background_color_( Ogre::ColourValue::Black ) // matches the default of Ogre::Viewport.
-  , stereo_enabled_( false )
-  , rendering_stereo_( false )
-  , left_camera_( nullptr )
-  , right_camera_( nullptr )
-  , right_viewport_( nullptr )
+  : QWidget( parent )
+  , render_system_( RenderSystem::get() )
 {
-  render_window_->setVisible(true);
-  render_window_->setAutoUpdated(true);
+  setAttribute(Qt::WA_OpaquePaintEvent,true);
+  setAttribute(Qt::WA_PaintOnScreen,true);
 
-  viewport_ = render_window_->addViewport( camera_ );
-  viewport_->setOverlaysEnabled( overlays_enabled_ );
-  viewport_->setBackgroundColour( background_color_ );
+#if QT_VERSION < QT_VERSION_CHECK(5, 0, 0)
+  // It is not clear to me why, but having this frame sub-widget
+  // inside the main widget makes an important difference (under X at
+  // least).  Without the frame and using this widget's winId()
+  // below causes trouble when using RenderWidget as a child
+  // widget.  The frame graphics are completely covered up by the 3D
+  // render, so using it does not affect the appearance at all.
+  this->renderFrame = new QFrame;
+  this->renderFrame->setLineWidth(1);
+  this->renderFrame->setFrameShadow(QFrame::Sunken);
+  this->renderFrame->setFrameShape(QFrame::Box);
+  this->renderFrame->show();
 
-#if OGRE_STEREO_ENABLE
-  viewport_->setDrawBuffer(Ogre::CBT_BACK);
+  QVBoxLayout *mainLayout = new QVBoxLayout;
+  mainLayout->setContentsMargins( 0, 0, 0, 0 );
+  mainLayout->addWidget(this->renderFrame);
+  this->setLayout(mainLayout);
 #endif
-  enableStereo(true);
 
-  setCameraAspectRatio();
+#if QT_VERSION < QT_VERSION_CHECK(5, 0, 0)
+  rviz::RenderSystem::WindowIDType win_id = this->renderFrame->winId();
+#else
+  rviz::RenderSystem::WindowIDType win_id = this->winId();
+#endif
+  QApplication::flush();
+
+#if QT_VERSION < QT_VERSION_CHECK(5, 0, 0)
+  QApplication::syncX();
+  double pixel_ratio = 1.0;
+#else
+  QWindow* window = windowHandle();
+  double pixel_ratio = window ? window->devicePixelRatio() : 1.0;
+#endif
+  render_window_ = render_system_->makeRenderWindow(win_id, static_cast<quint32>(width()), static_cast<quint32>(height()), pixel_ratio);
+
+  OgreViewportSupport::initialize();
 }
 
 QtWidgetOgreRenderWindow::~QtWidgetOgreRenderWindow()
 {
-  enableStereo(false);  // free stereo resources
-}
-
-//------------------------------------------------------------------------------
-bool QtWidgetOgreRenderWindow::enableStereo (bool enable)
-{
-  bool was_enabled = stereo_enabled_;
-  stereo_enabled_ = enable;
-  setupStereo();
-  return was_enabled;
-}
-
-void QtWidgetOgreRenderWindow::setupStereo()
-{
-  bool render_stereo = stereo_enabled_ && RenderSystem::get()->isStereoSupported();
-
-  if (render_stereo == rendering_stereo_)
-    return;
-
-  rendering_stereo_ = render_stereo;
-
-  if (rendering_stereo_)
-  {
-    right_viewport_ = render_window_->addViewport( NULL, 1 );
-#if OGRE_STEREO_ENABLE
-    right_viewport_->setDrawBuffer(Ogre::CBT_BACK_RIGHT);
-    viewport_->setDrawBuffer(Ogre::CBT_BACK_LEFT);
-#endif
-
-    setOverlaysEnabled(overlays_enabled_);
-    setBackgroundColor(background_color_);
-    if (camera_)
-      setCamera(camera_);
-
-    // addListener causes preViewportUpdate() to be called when rendering.
-    render_window_->addListener(this);
-  }
-  else
-  {
-    render_window_->removeListener(this);
-    render_window_->removeViewport(1);
-    right_viewport_ = NULL;
-
-#if OGRE_STEREO_ENABLE
-    viewport_->setDrawBuffer(Ogre::CBT_BACK);
-#endif
-
-    if (left_camera_)
-      left_camera_->getSceneManager()->destroyCamera( left_camera_ );
-    left_camera_ = NULL;
-    if (right_camera_)
-      right_camera_->getSceneManager()->destroyCamera( right_camera_ );
-    right_camera_ = NULL;
-  }
-
-}
-
-// this is called just before rendering either viewport when stereo is enabled.
-void QtWidgetOgreRenderWindow::preViewportUpdate(
-      const Ogre::RenderTargetViewportEvent& evt)
-{
-  Ogre::Viewport* viewport = evt.source;
-
-  const Ogre::Vector2& offset = camera_->getFrustumOffset();
-  const Ogre::Vector3 pos = camera_->getPosition();
-  const Ogre::Vector3 right = camera_->getRight();
-  const Ogre::Vector3 up = camera_->getUp();
-
-  if (viewport == right_viewport_)
-  {
-    if (camera_->getProjectionType() != Ogre::PT_PERSPECTIVE || !right_camera_)
-    {
-      viewport->setCamera( camera_ );
-      return;
-    }
-
-    Ogre::Vector3 newpos = pos
-                           + right * offset.x
-                           + up * offset.y;
-
-    right_camera_->synchroniseBaseSettingsWith(camera_);
-    right_camera_->setFrustumOffset(-offset);
-    right_camera_->setPosition(newpos);
-    viewport->setCamera(right_camera_);
-  }
-  else if (viewport == viewport_)
-  {
-    if (camera_->getProjectionType() != Ogre::PT_PERSPECTIVE || !left_camera_)
-    {
-      viewport->setCamera( camera_ );
-      return;
-    }
-
-    Ogre::Vector3 newpos = pos
-                           - right * offset.x
-                           - up * offset.y;
-
-    left_camera_->synchroniseBaseSettingsWith(camera_);
-    left_camera_->setFrustumOffset(offset);
-    left_camera_->setPosition(newpos);
-    viewport->setCamera(left_camera_);
-  }
-  else
-  {
-    ROS_WARN("Begin rendering to unknown viewport.");
-  }
-}
-
-void QtWidgetOgreRenderWindow::postViewportUpdate(
-      const Ogre::RenderTargetViewportEvent& evt)
-{
-  Ogre::Viewport* viewport = evt.source;
-
-  if (viewport == right_viewport_)
-  {
-    // nothing to do here
-  }
-  else if (viewport == viewport_)
-  {
-    viewport->setCamera(camera_);
-  }
-  else
-  {
-    ROS_WARN("End rendering to unknown viewport.");
-  }
-
-  if(!right_camera_->isCustomProjectionMatrixEnabled()) {
-    right_camera_->synchroniseBaseSettingsWith(camera_);
-    right_camera_->setFrustumOffset(-camera_->getFrustumOffset());
-  }
-  right_viewport_->setCamera(right_camera_);
-}
-
-Ogre::Viewport* QtWidgetOgreRenderWindow::getViewport () const
-{
-  return viewport_;
-}
-
-void QtWidgetOgreRenderWindow::setCamera( Ogre::Camera* camera )
-{
-  if (camera)
-  {
-    camera_ = camera;
-    viewport_->setCamera( camera );
-
-    setCameraAspectRatio();
-
-    if (camera_ && rendering_stereo_ && !left_camera_)
-    {
-      left_camera_ = camera_->getSceneManager()->createCamera(camera_->getName() + "-left");
-    }
-    if (camera_ && rendering_stereo_ && !right_camera_)
-    {
-      right_camera_ = camera_->getSceneManager()->createCamera(camera_->getName() + "-right");
-    }
-
-    update();
-  }
-}
-
-void QtWidgetOgreRenderWindow::setOverlaysEnabled( bool overlays_enabled )
-{
-  overlays_enabled_ = overlays_enabled;
-  viewport_->setOverlaysEnabled( overlays_enabled );
-  if (right_viewport_)
-  {
-    right_viewport_->setOverlaysEnabled( overlays_enabled );
-  }
-}
-
-void QtWidgetOgreRenderWindow::setBackgroundColor( Ogre::ColourValue background_color )
-{
-  background_color_ = background_color;
-  viewport_->setBackgroundColour( background_color );
-  if (right_viewport_)
-  {
-    right_viewport_->setBackgroundColour( background_color );
-  }
 }
 
 void QtWidgetOgreRenderWindow::setFocus( Qt::FocusReason reason )
@@ -272,42 +112,19 @@ QRect QtWidgetOgreRenderWindow::rect() const
   return QWidget::rect();
 }
 
-void QtWidgetOgreRenderWindow::setCameraAspectRatio()
+void QtWidgetOgreRenderWindow::updateScene()
 {
-  if ( camera_ )
+  QWidget::update();
+}
+
+void QtWidgetOgreRenderWindow::moveEvent(QMoveEvent *event)
+{
+  QWidget::moveEvent(event);
+
+  if(event->isAccepted() && render_window_)
   {
-    camera_->setAspectRatio( Ogre::Real( width() ) / Ogre::Real( height() ) );
-    if (right_camera_ ) {
-      right_camera_->setAspectRatio( Ogre::Real( width() ) / Ogre::Real( height() ) );
-    }
-
-    if ( camera_->getProjectionType() == Ogre::PT_ORTHOGRAPHIC )
-    {
-      Ogre::Matrix4 proj;
-      buildScaledOrthoMatrix( proj,
-                              -width() / ortho_scale_ / 2, width() / ortho_scale_ / 2,
-                              -height() / ortho_scale_ / 2, height() / ortho_scale_ / 2,
-                              camera_->getNearClipDistance(), camera_->getFarClipDistance() );
-      camera_->setCustomProjectionMatrix(true, proj);
-    }
+    render_window_->windowMovedOrResized();
   }
-}
-
-void QtWidgetOgreRenderWindow::setOrthoScale( float scale )
-{
-  ortho_scale_ = scale;
-
-  setCameraAspectRatio();
-}
-
-void QtWidgetOgreRenderWindow::setPreRenderCallback( boost::function<void ()> func )
-{
-  pre_render_callback_ = func;
-}
-
-void QtWidgetOgreRenderWindow::setPostRenderCallback( boost::function<void ()> func )
-{
-  post_render_callback_ = func;
 }
 
 //------------------------------------------------------------------------------
@@ -339,9 +156,17 @@ void QtWidgetOgreRenderWindow::paintEvent( QPaintEvent* )
 }
 
 //------------------------------------------------------------------------------
-void QtWidgetOgreRenderWindow::resizeEvent( QResizeEvent* event )
+void QtWidgetOgreRenderWindow::resizeEvent( QResizeEvent* )
 {
-  RenderWidget::resizeEvent( event );
+  if( render_window_ )
+  {
+    // render_window_->writeContentsToFile() (used in
+    // VisualizationFrame::onSaveImage()) does not work right for
+    // window with an odd width, so here I just always force it to be
+    // even.
+    render_window_->resize( static_cast<quint32>(width() + (width() % 2)), static_cast<quint32>(height()) );
+    render_window_->windowMovedOrResized();
+  }
 
   if( render_window_ )
   {
@@ -349,7 +174,7 @@ void QtWidgetOgreRenderWindow::resizeEvent( QResizeEvent* event )
 
     if( auto_render_ )
     {
-      update();
+      updateScene();
     }
   }
 }
