+++ src/rviz/render_panel.cpp
@@ -194,7 +194,7 @@ void RenderPanel::setViewController( ViewController* controller )
   }
   else
   {
-    render_window_->setCamera( NULL );
+    render_window_->setCamera( nullptr );
   }
 }
 
@@ -204,7 +204,10 @@ void RenderPanel::showContextMenu( boost::shared_ptr<QMenu> menu )
   context_menu_ = menu;
   context_menu_visible_ = true;
 
-  QApplication::postEvent( this, new QContextMenuEvent( QContextMenuEvent::Mouse, QPoint() ));
+  QWidget* widget = dynamic_cast<QWidget*>(render_window_);
+  if (widget) {
+    QApplication::postEvent( widget, new QContextMenuEvent( QContextMenuEvent::Mouse, QPoint() ));
+  }
 }
 
 void RenderPanel::onContextMenuHide()
